Move Achievement implementation to single file
